// src/webparts/first/components/types.ts
 export interface IUser  {
    Id?: string;
    first_name?: string;
    last_name?: string;
    email?: string;
    gender?: string;
    designation?: string;
    phone_no?: string;
    city?: string;
    department?: string;
    date_of_birth?: string;
    language?: string;
    Address?:any;
    Title?:any;
    imagePath?:any
  }
  export interface UserE{
    Id?:string;
    first_name?: string;
    email?: string;
    designation?: string;
    ImageUrl?:string;
  }
  